+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - Resumes replay when the app becomes active (#4303)
 - Session replay redact view with transformation (#4308)
 - Don't redact clipped views (#4325)
+- Correct redact UIView with higher zPosition (#4309) 
+- Don't redact clipped views (#4325)
+- Session replay for crash not created because of a race condition (#4314)
 - Double-quoted include, expected angle-bracketed instead (#4298)
 - Stop using `redactAllText` as an indicator tha redact is enabled (#4327)